Re-add Guild intent to get channel details
Channel had "disappeared" from the VoiceState objects passed to the
onVoiceStateChange event handler because I removed the intent to get
guild details while messing around with a prior implementation. This
adds it back so we can inspect channel details to decide what to write
in our greeting message.

This change also attempts to modularize the code a little better by
breaking in into a couple functions and passing values around rather
than relying on any global state. With some luck, I got them all, but
there might still be a few in there.