hellobirb: Enhance greeting message, add typing indicator
Make the greeting a little more fun with emoji and some other tidbits
that may appear if the stars line up. Also add a typing indicator so it
looks like bot is actually typing a message to users. Also set an
activity (the Discord user status message that appears beneath the
username in the user list) when the bot is ready.

Add a README to make the landing page look a little nicer.