hellobirb: ESLint fixes
Ran ESLint manually because of course the VSCode extension doesn't
automatically run it for some reason I have yet to figure out. And these
ESLint settings are probably conflicting with Prettier, so I should
figure out how to get these to coexist peacefully. Do people still use
Prettier, or does ESLint do enough on its own now with the --fix flag?
Who can keep up?

Once again, as it has been before, we see that it's not about the code
you write, but all the things you must do to juggle it and get it
prepped and delivered to where you think it belongs.