hellobirb: Add weather, weekend prompt, other small things
This change looks substantial because eslint/prettier reformatted
everything, so maybe ignore whitespace if you need to see what really
changed and this might help a bit.

Add a weather module to pull weather for a globally-configured location
and report the weather when a user joins. The message doesn't mention
the location, so if multiple locations are ever supported, this will
need to be considered.

Add weekend prompting to encourage some converation. E.g., "How was your
weekend?" on Mondays.

Ensure that "You're the first one here" and associated content is only
shown once a day, even if everyone leaves the room and someone comes
back on the same day. This is done by tracking the time of the most
recent greeting in the "database" and only adding the "first" content if
the most recent greeting isn't the current day. The database is migrated
from the old format on startup, and once the migration has run, this
code should be removed.

Add a prompt to show up occasionally reminding the user to drink water.

Add 4-Oct National Taco Day to special day reminders.