hellobirb: Address date issues
Was using date-fns Duration types to determine how long it'd been since
seeing a user, and looking at the "days" field. But if it's been, say,
exactly a week, then days will be 0. I.e., days is not a measure of the
duration in days, just the number of days that don't fit into larger
buckets. Fortunately, date-fns has other functions that reliably return
the number of days between dates.

Handle a failure case for sendTyping which now seems to be returning 403
Forbidden when called on the instance on my test server. It wasn't doing
this as recently as last night. May follow up on the Discord developers
server.

Remove the temporary "database" migration code now that it's been run in
"prod."

A few other date-fns were added in to use the library more fully and
replace the homegrown equivalents I was using, simple as they may have
been.

Users who haven't been seen before now get an "introduction."

Tweaked the ordering of some of the messages in the greeting to make it
flow a little better.

Added more shows that hellobirb can be watching.