hellobirb: Fix weather emoji and add tests
Prevent the emoji substituion in the weather forecast from jacking up
the text. Also add a few more conditions and vairations on existing
conditions. Add a test to ensure it works as expected.

Clarify language about when weather will be shown to reflect the actual
behavior that was taking place, since we were only tracking the last
time any greeting was issued, not the last time the weather was shared.
This could be done, but involves some more complicated closures or some
sort of persistent storage, so the easiest thing to do is to just not
share weather unless we haven't greeted anyone for however long, rather
than worrying about when the weather was last shared. This is totally
fine.

Also tweak an ESLint configuration to allow ES 2020 features, specifally
optional chaining here.