hellobirb: Change order of greeting messages
Put weather last to make the other occasional bits like MOTD,
on-this-day, and post-weekend greetings more prominent, since a user may
"tune out" the weather and ignore anything past it, even if it's a
good message.

Fix the count of active users in the channel by ignoring users who don't
have video enabled: we were getting bogus reports of being only two away
from "Hollywood Squares" even though fewer than 7 people had video
enabled.

Add a couple alternate suffix messages when multiple gifts are given.

Make logs easier to read by using user's displayName rather than
member ID.

Add a dev mode option to force the weather always to be reported, and
lengthen the required gap between weather reports from two to three
hours.

Also not reflected in this change, but something that prompted this
session in the first place was the repeated inability for the bot to
assign the presence role in the "production" guild, even though it could
do it in my test guild. The problem, silly enough, was that the bot's
role (hellobirb) was lower in the list of roles than the role it was
trying to add and remove. Physically dragging the bot's role above the
presence role it was trying to add and remove allowed the bot to
successfully add and remove it. Derp.