hellobirb: Only reply to messages in watch-channel guild
If more than one instance of hellobirb were running, they would all
receive messages from any guild an instance was watching, and possibly
reply (and presumably react with emoji, though this behavior was not
witnessed) to any message that satisfied the conditions for replying,
no matter what guild it was in. This had the practical effect of getting
multiple replies for a single mention. E.g., a user thanking hellobirb
would receive replies from every copy of hellobirb that was running.
This has been fixed by making sure the channel where the message was
received is contained in the guild's channel collection.

Also, if a user added certain punctuation after a mention of the bot
user, (e.g., "@hellobirb!"), it would show up as a _role_ mention in the
mention object inside the message payload, rather than a user mention
(since bot users also have a role with the same name, by default). This
meant that any mention with an exclamation point or a period, for
instance, would not be identified as something the bot should respond
to. This has been fixed by checking to see whether any roles mentioned
in the message payload are the bot's auto-assigned role.

Additionally, a few extra reactions have been added for extra
interactivity.