hellobirb: Mute unmuted users on channel join
If a user joins the channel and is unmuted, mute them automatically. If
the channel is configured properly for silence, this will only affect
administrators or other specially-privileged users that have the ability
to speak in a channel. The user will be server-muted, which they cannot
undo themselves, so the bot also watches for server mutes when users
leave the channel and remotes the server mute. This could potentially be
incorrect ia a server mute was applied manually, but for the use case
here, this is is not a concern.

Also note that the mute is not applied until the user enables their
camera, which is technically not correct since a user could come in
without camera, unmuted. But since the bot is only looking for the case
when a user turns on their camera, and this feature is unlikely to have
too great an effect, it's not worth complicating things with another
state right now. Worst case is that a user could turn off their camera
and remain in the room unmuted, which could be bad, but this is
unlikely.

This change also adds some more gifts and things for birb to watch.