hellobirb: Support multiple messages of the day
Glob for motd*.txt to determine messages for the day. This allows an
operator to simulate a sort of "stream of consciousness" by the birb,
where it keeps adding over a period of time.  The files may start with #
DELAY: <delay-in-milliseconds>\n\n to affect a custom delay between
messages, where the delay indicates how long will be waited before the
contents of the file that's annotated are sent (minus the delay
instruction, of course). If there is no delay instruction, the default
typing delay is used.

Speaking of the default typing delay, this change makes it an explicit
configuration value, defaulting to 3000 milliseconds. This means that
the "alwaysGreet" devMode option will not affect typing delay as it did
previously, where it eliminated delays when this devMode flag was set.
This was a conflation of concerns made clear by wanting to exercise the
"conversational" manner described above by the custom delays.

Also add a couple more things to watch and another respose when birb is
thanked.