hellobirb: Use longer delays before on-this-day and MOTD
Make hellobirb wait 5+ minutes for MOTDs, and 10+ minutes before
on-this-day messages are sent, in an attempt to make them seem more
"organically" sent. It's kind of bullshit, but it might change the feel
if people are actually around and starting to interact with one another.
It may increase enagaement with other members (or birb itself) or help keep
momentum going if it's already going.

Also: break out the sending of the on-this-day message much like how
sendMotd works, though a little simpler, because on-this-day does not
currently support multiple messages—though maybe if it did, the sendMotd
function could pretty much just be used for both, although log messages
would need to be simplified or something else would need to be done to
accurately log what action is being performed.

A few minor updates to the 12-01.txt on-this-day (OTD).