hellobirb: Support embeds for on-this-day
Send on-this-day messages as embed messages. The message files in the
on-this-day directory can remain text files, or can be JSON files to
directly specify the contents of the embed. Note that with this change,
images cannot be linked directly into the body of the message anymore;
they will only appear as links if this is done. They must be specified
as the image.url property in the embed. So, if an on-this-day message is
to have an image, the message will need to be saved as a JSON file in
the on-this-day message.